Cleaned Pushover
Pushover is now working to send notifications that bypass each users quiet hours.
However, until a feature request is implemented, we cannot acknowledge any messages sent to a group of users.
Each individual user would have to acknowledge every message in order to Pushover to set the status of the
message to be acknowledged.

Signed-off-by: Jake Champlin <jchamplin@thedatacave.com>